# new_project/contracts/openapi/investment-api.yaml
openapi: 3.1.0
info:
  title: Personal Investment API
  version: 0.2.0
  description: 本地开发切片。所有金额以原币种两位小数最小单位的十进制字符串传输，绝不使用浮点数。
servers:
  - url: http://127.0.0.1:<API_PORT>
paths:
...
  /api/v1/market/instruments:
    get:
      summary: 分页查询全局行情标的和显式 Tushare 代码映射
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: cursor
          in: query
          schema: {type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'}
      responses:
        '200':
          description: 稳定游标页
          content:
            application/json:
              schema: {$ref: '#/components/schemas/InstrumentPage'}
    post:
      summary: 创建或复用全局标的主数据
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstrumentRequest'
      responses:
        '201':
          description: 已创建；同自然键同定义或同幂等键返回首个响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Instrument'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /api/v1/market/instruments/{instrumentId}/source-codes/TUSHARE_PRO:
    put:
      summary: 维护标的的显式 Tushare Pro 代码
      security:
        - bearerAuth: []
      parameters:
        - name: instrumentId
          in: path
          required: true
          schema: {type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/TushareCodeRequest'}
      responses:
        '200':
          description: 代码已更新
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Instrument'}
...
  /api/v1/ledger/snapshots:
    get:
      summary: 列出当前 owner 的加密账本快照元数据
      description: 返回只读元数据和内容 SHA-256；快照正文必须通过 owner 认证下载，永不暴露对象存储地址。
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100, default: 20}
      responses:
        '200':
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LedgerSnapshotListResponse'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
    post:
      summary: 为当前不可变账本生成或复用当日版本快照
      description: 快照正文以服务端加密对象保存，元数据及生成操作写入审计日志。空账本不能创建快照。
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 128}
      responses:
        '201':
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LedgerSnapshot'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
  /api/v1/ledger/snapshots/{ledgerSnapshotId}/download:
    get:
      summary: 下载已校验的 owner 快照 JSON
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerSnapshotId'
      responses:
        '200':
          description: 附件流；响应头包含 snapshot ID 与 SHA-256。
          content:
            application/json: {schema: {type: string, format: binary}}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '404': {$ref: '#/components/responses/Problem'}
  /api/v1/ledger/snapshots/{ledgerSnapshotId}/restore:
    post:
      summary: 在空账本工作区原子恢复快照
      description: 服务端先锁住 owner 的账本状态，再确认没有任何账户或账务事实。只会以新的 IMPORT 事实追加恢复；已有账本一律拒绝，不存在覆盖、删除或回滚接口。恢复前必须验证快照 SHA-256 与所有引用标的的主数据。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LedgerSnapshotId'
        - name: Idempotency-Key
          in: header
          required: true
          schema: {type: string, minLength: 1, maxLength: 128}
      responses:
        '201':
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LedgerSnapshotRestoreResponse'}
        '400': {$ref: '#/components/responses/Problem'}
        '401': {$ref: '#/components/responses/Problem'}
        '409': {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    LedgerSnapshotId:
      name: ledgerSnapshotId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
  responses:
    Problem:
      description: RFC 7807 风格错误响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    CreateInstrumentRequest:
      type: object
      required: [market, exchange, symbol, displayName, assetType, nativeCurrency]
      properties:
        market:
          type: string
          maxLength: 32
        exchange:
          type: string
          maxLength: 32
        symbol:
          type: string
          maxLength: 64
        displayName:
          type: string
          maxLength: 256
        assetType:
          type: string
          enum: [EQUITY, ETF, INDEX, FUTURE, OPTION]
        nativeCurrency:
          type: string
          enum: [CNY, USD]
        maturityDate:
          type: [string, 'null']
          format: date
        future:
          oneOf:
            - $ref: '#/components/schemas/FutureSpecification'
            - type: 'null'
        option:
          oneOf:
            - $ref: '#/components/schemas/OptionSpecification'
            - type: 'null'
        tushareCode:
          type: [string, 'null']
          maxLength: 64
        underlyingInstrumentId:
          type: [string, 'null']
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
    FutureSpecification:
      type: object
      required: [productCode, contractMultiplierCent]
      properties:
        productCode:
          type: string
          enum: [IC, IM]
        contractMultiplierCent:
          type: string
          pattern: '^[1-9][0-9]*$'
    OptionSpecification:
      type: object
      required: [underlyingInstrumentId, optionRight, strikePriceCent, contractMultiplier]
      properties:
        underlyingInstrumentId:
          type: string
          pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
        optionRight:
          type: string
          enum: [PUT, CALL]
        strikePriceCent:
          type: string
          pattern: '^[1-9][0-9]*$'
        contractMultiplier:
          type: string
          pattern: '^[1-9][0-9]*$'
    Instrument:
      allOf:
        - $ref: '#/components/schemas/CreateInstrumentRequest'
        - type: object
          required: [instrumentId, status]
          properties:
            instrumentId:
              type: string
              pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'
            status:
              type: string
              enum: [ACTIVE]
    InstrumentPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/Instrument'}
        nextCursor:
          type: [string, 'null']
    TushareCodeRequest:
      type: object
      required: [tushareCode]
      properties:
        tushareCode:
          type: string
          minLength: 1
          maxLength: 64
...
    LedgerSnapshot:
      type: object
      required: [ledgerSnapshotId, asOfDate, sourceLedgerVersion, contentSha256Hex, createdAt]
      properties:
        ledgerSnapshotId: {type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'}
        asOfDate: {type: string, format: date}
        sourceLedgerVersion: {type: string, pattern: '^[1-9][0-9]*$'}
        contentSha256Hex: {type: string, pattern: '^[a-f0-9]{64}$'}
        createdAt: {type: string, format: date-time}
    LedgerSnapshotListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/LedgerSnapshot'}
    LedgerSnapshotRestoreResponse:
      type: object
      required: [ledgerSnapshotId, restoredAccountCount, restoredTransactionCount, targetLedgerVersion]
      properties:
        ledgerSnapshotId: {type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$'}
        restoredAccountCount: {type: integer, minimum: 1}
        restoredTransactionCount: {type: integer, minimum: 1}
        targetLedgerVersion: {type: string, pattern: '^[1-9][0-9]*$'}
...
    Problem:
      type: object
      required: [code, message, traceId, details]
      properties:
        code:
          type: string
        message:
          type: string
        traceId:
          type: string
        details:
          type: array
          items:
            type: string